@@ -3,14 +3,22 @@
 
 package validation
 
+type Document struct {
+	// Ajoutez les champs pertinents pour un document ici
+	// Par exemple:
+	ID      string
+	Content string
+	IsValid bool
+}
+
 type ValidationIssue struct {
-Type    string
-Message string
-Fixable bool
+	Type    string
+	Message string
+	Fixable bool
 }
 
 type ValidationReport struct {
-Issues    []ValidationIssue
-Summary   string
-Validated bool
+	Issues    []ValidationIssue
+	Summary   string
+	Validated bool
 }
