@@ -1,45 +1,51 @@
-// Tests unitaires pour le package scanmodules.
+// core/scanmodules/scanmodules_test.go
 package scanmodules
 
 import (
 	"os"
-	"path/filepath"
 	"testing"
 )
 
-// Test de la fonction DetectLang.
-func TestDetectLang(t *testing.T) {
-	tests := []struct {
-		filename string
-		expected string
-	}{
-		{"main.go", "Go"},
-		{"script.js", "Node.js"},
-		{"module.py", "Python"},
-		{"README.md", "unknown"},
+// TestMain est une fonction spéciale qui s'exécute avant tous les tests du package.
+// Elle permet de configurer l'environnement de test et de nettoyer après l'exécution des tests.
+func TestMain(m *testing.M) {
+	// Créer un répertoire temporaire pour les fichiers de sortie
+	tempDir, err := os.MkdirTemp("", "test_scanmodules")
+	if err != nil {
+		panic(err)
 	}
+	defer os.RemoveAll(tempDir) // Nettoyer après les tests
 
-	for _, tt := range tests {
-		result := DetectLang(tt.filename)
-		if result != tt.expected {
-			t.Errorf("DetectLang(%q) = %q; want %q", tt.filename, result, tt.expected)
-		}
+	// Changer le répertoire de travail pour le répertoire temporaire
+	originalDir, err := os.Getwd()
+	if err != nil {
+		panic(err)
 	}
+	err = os.Chdir(tempDir)
+	if err != nil {
+		panic(err)
+	}
+	defer os.Chdir(originalDir) // Revenir au répertoire d'origine après les tests
+
+	// Exécuter les tests
+	exitCode := m.Run()
+
+	os.Exit(exitCode)
 }
 
-// Test de la fonction ScanDir sur un dossier temporaire.
-func TestScanDir(t *testing.T) {
-	tmpDir := t.TempDir()
-	files := []string{"a.go", "b.js", "c.py", "d.txt"}
-	for _, f := range files {
-		os.WriteFile(filepath.Join(tmpDir, f), []byte("test"), 0644)
+func TestRunScanModules(t *testing.T) {
+	// Exécuter la fonction RunScanModules
+	err := RunScanModules()
+	if err != nil {
+		t.Fatalf("RunScanModules a retourné une erreur: %v", err)
 	}
 
-	modules, err := ScanDir(tmpDir)
-	if err != nil {
-		t.Fatalf("ScanDir error: %v", err)
-	}
-	if len(modules) != len(files) {
-		t.Errorf("ScanDir found %d modules; want %d", len(modules), len(files))
+	// Vérifier que les fichiers de sortie existent
+	files := []string{"arborescence.txt", "modules.txt", "modules.json"}
+	for _, file := range files {
+		_, err := os.Stat(file)
+		if os.IsNotExist(err) {
+			t.Errorf("Le fichier %s n'a pas été créé", file)
+		}
 	}
 }
