@@ -7,9 +7,9 @@ import (
 	"path/filepath"
 	"time"
 
-	"github.com/gerivdb/email-sender-1/tools/core/toolkit"
-	"github.com/gerivdb/email-sender-1/tools/operations/validation"
-	toolkitpkg "github.com/gerivdb/email-sender-1/tools/pkg/manager"
+	"email_sender/tools/core/toolkit"
+	"email_sender/tools/operations/validation"
+	toolkitpkg "email_sender/tools/pkg/manager"
 )
 
 // Simple standalone test runner for validation
