@@ -3,10 +3,10 @@ package main
 import (
 	"fmt"
 	/*
-	   "github.com/gerivdb/email-sender-1/tools/core/toolkit"
-	   "github.com/gerivdb/email-sender-1/tools/pkg/manager"
+	   "email_sender/tools/core/toolkit"
+	   "email_sender/tools/pkg/manager"
 	*/)
 
 func main() {
-fmt.Println("Erreur : modules privés 'github.com/gerivdb/email-sender-1' non disponibles. Corrigez les imports pour compiler ce test.")
+	fmt.Println("Erreur : modules privés 'email_sender' non disponibles. Corrigez les imports pour compiler ce test.")
 }
